+++ src/components/App/AppButtonBack.vue
@@ -1,0 +1,13 @@
+<template>
+  <AppButton
+    icon="mdi:arrow-left"
+    variant="link"
+    @click="$router.back()"
+  >
+    <slot />
+  </AppButton>
+</template>
+
+<script setup>
+import AppButton from '@/components/App/AppButton.vue'
+</script>