+++ src/components/App/Sidebar.vue
@@ -1,69 +1,56 @@
 <script setup>
-import router from '@/router'
+import { Icon } from '@iconify/vue'
+
+import AppNavMenuItem from '@/components/App/AppNavMenuItem.vue'
 
 defineOptions({
   name: 'AppSidebar',
 })
 
-const redirectTo = (routename) => {
-  router.push({ name: routename })
-}
 </script>
 
 <template>
-  <aside class="bg-primary-900 text-white p-4">
-    <ul class="space-y-2">
-      <li>
-        <div
-          class="flex items-center justify-between text-secondary px-4 py-2 hover:bg-gray-700 cursor-pointer"
-          @click="redirectTo('home')"
-        >
-          <div class="flex items-center">
-            <span>{{ $t('navigation.menu.home') }}</span>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div
-          class="flex items-center justify-between text-secondary px-4 py-2 hover:bg-gray-700 cursor-pointer"
-          @click="redirectTo('intro')"
-        >
-          <div class="flex items-center">
-            <span>{{ $t('navigation.menu.intro') }}</span>
-          </div>
-        </div>
-      </li>
-      <div class="border my-6" />
-      <li>
-        <div
-          class="flex items-center justify-between text-secondary px-4 py-2 hover:bg-gray-700 cursor-pointer"
-          @click="redirectTo('register-form')"
-        >
-          <div class="flex items-center">
-            <span>{{ $t('navigation.menu.challengeOne') }}</span>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div
-          class="flex items-center justify-between text-secondary px-4 py-2 hover:bg-gray-700 cursor-pointer"
-          @click="redirectTo('user-detail')"
-        >
-          <div class="flex items-center">
-            <span>{{ $t('navigation.menu.challengeTwo') }}</span>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div
-          class="flex items-center justify-between text-secondary px-4 py-2 hover:bg-gray-700 cursor-pointer"
-          @click="redirectTo('event-calendar')"
-        >
-          <div class="flex items-center">
-            <span>{{ $t('navigation.menu.challengeThree') }}</span>
-          </div>
-        </div>
-      </li>
-    </ul>
-  </aside>
+  <div
+    class="fixed bottom-0 pt-2 pb-1 shadow-md w-screen flex justify-around lg:justify-start lg:items-start lg:relative lg:flex-col gap-4 lg:w-48 border-t border-secondary-800 bg-gray-800 text-wsmoke lg:py-8 px-6 z-10"
+  >
+    <AppNavMenuItem :to="{ name: 'introduction' }">
+      <template #icon>
+        <Icon
+          icon="mdi:home"
+          class="text-2xl lg:hidden"
+        />
+      </template>
+      <template #label>
+        <p class="text-sm lg:text-base">
+          {{ $t('navigation.navbar.intro') }}
+        </p>
+      </template>
+    </AppNavMenuItem>
+    <AppNavMenuItem :to="{ name: 'calendar' }">
+      <template #icon>
+        <Icon
+          icon="mdi:calendar"
+          class="text-2xl lg:hidden"
+        />
+      </template>
+      <template #label>
+        <p class="text-sm lg:text-base">
+          {{ $t('navigation.navbar.calendar') }}
+        </p>
+      </template>
+    </AppNavMenuItem>
+    <AppNavMenuItem :to="{ name: 'profile' }">
+      <template #icon>
+        <Icon
+          icon="mdi:account"
+          class="text-2xl lg:hidden"
+        />
+      </template>
+      <template #label>
+        <p class="text-sm lg:text-base">
+          {{ $t('navigation.navbar.profile') }}
+        </p>
+      </template>
+    </AppNavMenuItem>
+  </div>
 </template>