+++ src/layouts/PrivateLayout.vue
@@ -8,11 +8,11 @@ defineOptions({
 </script>
 
 <template>
-  <div class="h-screen w-screen flex flex-col">
+  <div class="h-screen w-screen flex flex-col bg-wsmoke">
     <AppNavbar />
     <div class="overflow-y-auto flex flex-1">
       <AppSidebar />
-      <main class="flex-1 overflow-y-auto">
+      <main class="flex-1 overflow-y-auto pb-16 lg:pb-8">
         <slot />
       </main>
     </div>