+++ src/views/UserDetailView.vue
@@ -1,5 +1,17 @@
 <script setup>
-import AppCard from '@/components/App/AppCard.vue'
+
+import { useWindowSize } from '@vueuse/core'
+import { computed, ref, watch } from 'vue'
+
+import AppAccordion from '@/components/App/AppAccordion.vue'
+import AppButton from '@/components/App/AppButton.vue'
+import AppButtonBack from '@/components/App/AppButtonBack.vue'
+import AppProfileUserImage from '@/components/App/AppProfileUserImage.vue'
+import BankInfoForm from '@/components/Forms/BankInfoForm.vue'
+import BasicUserInfoForm from '@/components/Forms/BasicUserInfoForm.vue'
+import CompanyInfoForm from '@/components/Forms/CompanyInfoForm.vue'
+import SecurityAndPrivacityInfoForm from '@/components/Forms/SecurityAndPrivacityInfoForm.vue'
+
 /*
 Reto 2: Vista de Datos del Usuario.
 
@@ -70,38 +82,126 @@ const userData = {
   ein: '20-9487066',
   ssn: '661-64-2976',
 }
+
+const { width } = useWindowSize()
+const isMobile = computed(() => width.value < 768)
+
+const tabs = [
+  {
+    id: 'basic-information',
+    translation: 'basicInformation',
+    data: userData,
+    component: BasicUserInfoForm,
+  },
+  {
+    id: 'company-information',
+    translation: 'company',
+    data: userData.company,
+    component: CompanyInfoForm,
+  },
+  {
+    id: 'bank-information',
+    translation: 'bankInformation',
+    data: userData.bank,
+    component: BankInfoForm,
+  },
+  {
+    id: 'security-and-privacity',
+    translation: 'privacity',
+    data: userData,
+    component: SecurityAndPrivacityInfoForm,
+  },
+]
+
+const activeTabIndex = ref('basic-information')
+const activeTabs = computed(() =>
+  isMobile.value
+    ? tabs
+    : tabs.filter(i => i.id === activeTabIndex.value),
+)
+
+watch(isMobile, () => {
+  activeTabIndex.value = tabs[0].id
+})
+
+const changeTab = (tab) => {
+  if (isMobile.value) return
+  activeTabIndex.value = tab.id
+}
+
 </script>
 
 <template>
-  <section class="h-full flex items-stretch text-white bg-gray-800">
-    <div class="lg:flex lg:flex-col lg:w-1/4 hidden border-x border-gray-700 bg-gray-800 bg-no-repeat bg-cover relative px-4 py-6">
-      <h3 class="text-wsmoke font-semibold text-2xl mb-4">
+  <section class="h-full flex items-stretch bg-wsmoke">
+    <!-- this menu will be hidden on small screens -->
+    <div
+      class="md:flex md:flex-col md:w-[40%] max-w-80 hidden border-x border-primary-200 bg-wsmoke bg-no-repeat text-primary-900 relative px-4 py-6"
+    >
+      <h3 class="font-semibold text-2xl mb-4">
         {{ $t('profile.navbar.title') }}
       </h3>
-      <p class="text-wsmoke">
+      <p>
         {{ $t('profile.navbar.subtitle') }}
       </p>
+
+      <div class="tab-container mt-10 hidden md:flex flex-col gap-2 items-start">
+        <AppButton
+          v-for="tab in tabs"
+          :key="`tab-${tab.translation}`"
+          variant="link"
+          :class="[ activeTabIndex === tab.id ? 'text-secondary-700' : '' ]"
+          @click="changeTab(tab)"
+        >
+          {{ $t(`profile.tabs.${tab.translation}`) }}
+        </AppButton>
+      </div>
     </div>
-    <div class="w-full flex items-center justify-center md:px-16 px-0 z-0 ">
-      <div
-        class="relative w-full max-w-[25rem] h-[-webkit-fill-available] mx-auto py-6 z-20  px-8 md:p-0 md:h-auto md:rounded md:shadow lg:shadow-none lg:rounded-none"
-      >
-        <AppCard>
-          <template #title>
-            <h2 class="text-lg font-semibold">
-              {{ userData.firstName }} {{ userData.lastName }}
-            </h2>
+    <!-- this content will show full screen in small screens -->
+    <div class="w-full flex flex-col md:flex-row">
+      <div class="flex items-center md:hidden my-6">
+        <AppButtonBack />
+        <h2
+          role="heading"
+          aria-level="2"
+          class="mb-0"
+        >
+          {{ $t('profile.navbar.title') }}
+        </h2>
+      </div>
+      <div class="accordion-container p-6 gap-4 flex flex-col w-full pb-20 lg:pb-8">
+        <AppProfileUserImage
+          :image-profile-url="userData.image"
+          :show-change-option="true"
+        />
+        <p class="text-primary-800 text-center font-medium text-xl">
+          {{ `${userData.firstName} ${userData.lastName}` }}
+        </p>
+        <AppAccordion
+          v-for="(tab, index) in activeTabs"
+          :id="`tab-${index}`"
+          :key="`tab-${tab.translation}`"
+          :default-open="index === 0 || activeTabIndex === tab.id"
+          :interaction="isMobile"
+        >
+          <template #header>
+            <h3
+              role="heading"
+              aria-level="3"
+              class="text-lg font-semibold mb-0"
+            >
+              {{ $t(`profile.tabs.${tab.translation}`) }}
+            </h3>
           </template>
-          <template #description>
-            <p>Email: {{ userData.email }}</p>
-            <p>Phone: {{ userData.phone }}</p>
-            <p>Age: {{ userData.age }}</p>
+          <template #content>
+            <component
+              :is="tab.component"
+              :data="tab.data"
+            >
+              <slot />
+            </component>
           </template>
-        </AppCard>
+        </AppAccordion>
       </div>
     </div>
   </section>
-  <pre>
-    {{ userData }}
-    </pre>
 </template>